src/requirements.txt
alembic==1.17.1
asyncpg==0.30.0
fastapi==0.120.4
pydantic-settings==2.11.0
SQLAlchemy==2.0.44
email-validator==2.3.0
uvicorn==0.38.0
pyjwt[crypto]==2.10.1
bcrypt==5.0.0
python-multipart==0.0.20